Fixes transform_dicts KeyError using Registry
Fixes the key error caused by transform_dicts() when the first record
of a dataset is missing one or more keys present in the remaining data.
Original code pulled the keys of the first record and used those to
determine which transforms to load. Changes pull the list of required
transforms from the table column Registry instead of incoming data